add a zoom reset button
