+++ static/app/views/performance/transactionSummary/transactionSpans/spanDetails/utils.tsx
@@ -47,3 +47,4 @@ export function spanDetailsRouteWithQuery({
 
 export const MIN = 'min';
 export const MAX = 'max';
+export const ZOOM_KEYS = [MIN, MAX];
